clean up db
